+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 #### 1.N.N - YYYY-MM-DD
 
 
+#### 1.1.1 - 2022-04-28
+
+- feat(DNSKEY): expanded algo from 1-5 to 1-16 (RFC 8624)
+    - warn if outside that range, not error
+
+
 #### 1.1.0 - 2022-04-22
 
 - feat(tinydns): add ipv4toOctal, octalToIPv4, base64toOctal, octalToBase64